requirements.txt
pillow
openai
python-dotenv
